LayoutTests/imported/w3c/web-platform-tests/dom/nodes/moveBefore/moveBefore-id-map-expected.txt


FAIL moveBefore() correctly updates id map assert_equals: expected Element node <span id="target" data-span=""></span> but got Element node <div id="target" data-target=""></div>
